Movement komplett
Movement mit kleiner Animation fertiggestellt. (packman teilbilder für Animation eingefügt)
FPS auf 60 gesenkt, damit packman richtig aungezeigt wird.

+++ src/EntitySystem/Entity.java
@@ -3,7 +3,7 @@ package EntitySystem;
 import java.awt.*;
 
 public abstract class Entity {
-    protected double posX,posY;
+    protected float posX,posY;
 
     public Entity(float posX, float posY){
         this.posX = posX;
